The compiler -> a compiler (#1226)
* The compiler -> a compiler

Search for instances of "the compiler" and where the text describes what a compliant compiler is is required or allowed to do, replace "the compiler" with "a compiler".  Where necessary, include "shall" or "may", as appropriate.

There are still instances of "the compiler" in the standard. Some are in types.md. Those are addressed in #1125.  Others are where the text describes "the compiler generated" backing field, class, or method.

* fix typo

* respond to feedback.

* Update standard/classes.md

Co-authored-by: Nigel-Ecma <6654683+Nigel-Ecma@users.noreply.github.com>

* Update standard/classes.md

Co-authored-by: Bill Wagner <wiwagn@microsoft.com>

---------

Co-authored-by: Rex Jaeschke <rex@RexJaeschke.com>
Co-authored-by: Jon Skeet <skeet@pobox.com>
Co-authored-by: Nigel-Ecma <6654683+Nigel-Ecma@users.noreply.github.com>